Everything

+++ nginx/swap2.sh
@@ -2,11 +2,18 @@
 # This shell is to swap from web1 to web2
 
 sed -i '' 's/image: activity/image: activity2/' docker-compose.yml
+sed -i '' 's/image: activity22/image: activity2/' docker-compose.yml
+
+docker-compose up
 
 var="$(docker ps | awk -F' ' '$2 ==  "activity"' | awk '{print $1}')"
 
-docker rm -f $var
+if [ -z "$var" ]
+then
+     echo "activity is not running"
+else
+      docker rm -f $var
+fi
 
-docker-compose up
 
 # service nginx reload
